+++ cardinal-android/app/src/main/res/values/strings.xml
@@ -192,5 +192,5 @@
     <string name="content_description_previous_page">Previous page</string>
     <string name="remove_as_default">Remove as default</string>
     <string name="set_as_default">Set as default</string>
-    <string name="add_save_places_and_they_ll_show_up_here">Add save places and they\'ll show up here.</string>
+    <string name="add_save_places_and_they_ll_show_up_here">Favorite places and they\'ll show up here.</string>
 </resources>
